refactor(cli): move cli components and hooks to cli module
