--- src/modules/vault/components/cli/index.ts
@@ -1,1 +1,0 @@
-export * from './CLISettingsCard';
